+++ plugins/tmdb.js
@@ -1,7 +1,7 @@
 tmdb = {
     metadata_xml: null,
 
-    init: function (metadata_xml) {
+    init: function (metadata_xml, server, type) {
         tmdb.insertTmdbLink(metadata_xml);
     },